requirements.txt
Flask==3.0.0
speedtest-cli==2.1.3
ping3==4.0.4
requests==2.31.0
gunicorn==21.2.0
